# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book sale",
        "description": "Простой сервис по пробаже книг",
        "version": "1"
    },
    "paths": {
...
        "/user/signup": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Регистрагия пользователя",
                "description": "Регистрагия пользователя",
                "operationId": "06a2a9fc67fdee725b42fdae9bae69b3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "Oleg"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "Oleg@mail.ru"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "23455"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "token": {
                                                        "type": "string",
                                                        "format": "JWT",
                                                        "example": "Fk32aopIeVO5KKgVu27FYwQMwE3"
                                                    },
                                                    "user": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "format": "int",
                                                                "example": "1"
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "format": "name",
                                                                "example": "Oleg"
                                                            },
                                                            "email": {
                                                                "type": "string",
                                                                "format": "email",
                                                                "example": "Oleg@mail.ru"
                                                            },
                                                            "email_verified_at": {
                                                                "type": "string",
                                                                "format": "TIMESTAMP",
                                                                "example": "2025-06-21T12:56:45.000000Z"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "TIMESTAMP",
                                                                "example": "2025-06-21T12:56:45.000000Z"
                                                            },
                                                            "updated_at": {
                                                                "type": "string",
                                                                "format": "TIMESTAMP",
                                                                "example": "2025-06-21T12:56:45.000000Z"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/signin": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Получение токена пользователя",
                "description": "Получение токена пользователя",
                "operationId": "250323879ca6c05172cbec88ed87bd38",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "Oleg@mail.ru"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "23455"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "token": {
                                                        "type": "string",
                                                        "format": "JWT",
                                                        "example": "Fk32aopIeVO5KKgVu27FYwQMwE3"
                                                    },
                                                    "user": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "format": "int",
                                                                "example": "1"
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "format": "name",
                                                                "example": "Oleg"
                                                            },
                                                            "email": {
                                                                "type": "string",
                                                                "format": "email",
                                                                "example": "Oleg@mail.ru"
                                                            },
                                                            "email_verified_at": {
                                                                "type": "string",
                                                                "format": "TIMESTAMP",
                                                                "example": "2025-06-21T12:56:45.000000Z"
                                                            },
                                                            "created_at": {
                                                                "type": "string",
                                                                "format": "TIMESTAMP",
                                                                "example": "2025-06-21T12:56:45.000000Z"
                                                            },
                                                            "updated_at": {
                                                                "type": "string",
                                                                "format": "TIMESTAMP",
                                                                "example": "2025-06-21T12:56:45.000000Z"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка"
                    }
                }
            }
        },
        "/user": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Поиск пользователя по имени",
                "description": "Поиск пользователя по имени",
                "operationId": "10551226e9bb607517f5e51dd4137ec2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "name",
                                        "example": "Oleg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Successfully"
                                        },
                                        "paginator": {
                                            "$ref": "#/components/schemas/Paginator"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "1"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "format": "name",
                                                        "example": "Oleg"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "format": "email",
                                                        "example": "Oleg@mail.ru"
                                                    },
                                                    "email_verified_at": {
                                                        "type": "string",
                                                        "format": "TIMESTAMP",
                                                        "example": "2025-06-21T12:56:45.000000Z"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "TIMESTAMP",
                                                        "example": "2025-06-21T12:56:45.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "TIMESTAMP",
                                                        "example": "2025-06-21T12:56:45.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Paginator": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "format": "int",
                        "example": "1"
                    },
                    "first_page_url": {
                        "type": "string",
                        "format": "url",
                        "example": "http://138.124.55.208/api/prod/chat?page=1"
                    },
                    "from": {
                        "type": "integer",
                        "format": "int",
                        "example": "1"
                    },
                    "last_page": {
                        "type": "integer",
                        "format": "int",
                        "example": "380"
                    },
                    "last_page_url": {
                        "type": "string",
                        "format": "url",
                        "example": "http://138.124.55.208/api/prod/chat?page=5"
                    },
                    "next_page_url": {
                        "type": "string",
                        "format": "url",
                        "example": "http://138.124.55.208/api/prod/chat?page=3"
                    },
                    "path": {
                        "type": "string",
                        "format": "url",
                        "example": "http://138.124.55.208/api/prod/chat?page=2"
                    },
                    "per_page": {
                        "type": "integer",
                        "format": "int",
                        "example": "30"
                    },
                    "to": {
                        "type": "integer",
                        "format": "int",
                        "example": "30"
                    },
                    "total": {
                        "type": "integer",
                        "format": "int",
                        "example": "11371"
                    }
                },
                "type": "object"
            }
        }
    },
...
}